clozr-engine/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
SQLAlchemy>=2.0
pydantic
python-dotenv
requests

